+++ doc/changes.rst
@@ -5,7 +5,19 @@ desitarget Change Log
 0.37.4 (unreleased)
 -------------------
 
-* No changes yet.
+* Extra columns and features in the random catalogs [`PR #606`_]:
+    * Better error messages and defaults for `bin/supplement_randoms`.
+    * Don't calculate APFLUX quantities if aprad=0 is passed.
+    * Pass the randoms through the `finalize` and `make_mtl` functions:
+        * To populate columns needed to run fiberassign on the randoms.
+        * Addresses `issue #597`_.
+    * Add the `BRICKID` column to the random catalogs.
+    * Also add a realistic `TARGETID` (and `RELEASE, BRICK_OBJID`).
+    * Recognize failure modes more quickly (and fail more quickly).
+    * Write out both "resolve" and "noresolve" (North/South) catalogs.
+
+.. _`issue #597`: https://github.com/desihub/desitarget/issues/597
+.. _`PR #606`: https://github.com/desihub/desitarget/pull/606
 
 0.37.3 (2020-04-15)
 -------------------